pyproject.toml
[tool.poetry]
name = "at-hackathon-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
Django = "^3.2.9"
djangorestframework-files = "^1.1.0"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-heroku = "^0.3.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.2"
dj-rest-auth = {extras = ["with_social"], version = "^2.1.11"}
djangorestframework-simplejwt = "^5.0.0"
django-s3-storage = "^0.13.5"
django-dotenv = "^1.4.2"
django-cors-headers = "^3.10.0"
Pillow = "^8.4.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest-black = "^0.3.12"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
